Safari-609.1.11.1/LayoutTests/inspector/indexeddb/requestDatabaseNames-expected.txt

CONSOLE MESSAGE: line 9: Created Database 'Database1'
CONSOLE MESSAGE: line 9: Created Database 'Database2'
CONSOLE MESSAGE: line 9: Created Database 'á‰'
CONSOLE MESSAGE: line 9: Created Database 'ğ†'

== Running test suite: IndexedDB.requestDatabaseNames
-- Running test case: ClearDatabases
-- Running test case: EnsureNoDatabases
PASS: No IndexedDB databases should exist initially.

-- Running test case: CreateDatabase1
PASS: A single IndexedDB database should exist.
["Database1"]

-- Running test case: CreateDatabase2
PASS: Two IndexedDB databases should exist.
["Database1","Database2"]

-- Running test case: CreateDatabase3
PASS: Three IndexedDB databases should exist.
["Database1","Database2","á‰"]

-- Running test case: CreateDatabase4
PASS: Four IndexedDB databases should exist.
["Database1","Database2","á‰","ğ†"]
